requirements.txt
anthropic>=0.42.0
google-cloud-bigquery>=3.25.0
google-auth>=2.30.0
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.8.0
openpyxl>=3.1.0
